alternate vcs: Fixed tests, cleaned code
