+++ src/towncrier/_novcs.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import os
 
 from typing import Container
